Add GitHub Actions workflow to package and publish project
This commit adds a new GitHub Actions workflow which will trigger on event 'release'. The workflow is configured to checkout the code, setup .NET environment, restore dependencies, build the project, run tests, package the build, and publish it onto NuGet using the specified API Key, all on an Ubuntu environment.