Update build and test commands in GitHub workflow
The build and test commands in the packageanddeploy.yml workflow file have been updated to remove specific paths. This should allow the commands to target the whole repository rather than a single project within it. Also, the PROJECT_FILE_PATH reference was removed in the publish_nuget step for more streamlining.