Update package metadata and improve deployment script
Updated the package metadata in the Boom.IdType.EntityFramework.csproj file to be more descriptive and accurate. The deployment script in packageanddeploy.yml has also been simplified and updated for better efficiency. This includes replacing old project references with the latest version and incorporating a new ID package update step.